Update aprxfeeder.py
Updated feeder script for improved error handling.

+++ aprxfeeder.py
@@ -1,13 +1,18 @@
 #!/usr/bin/python3
 
-import sys
+import sys, os
 
-try:
+if os.path.exists('/tmp/wxalerts.txt'):
     with open('/tmp/wxalerts.txt', 'r+') as f:
-        lines = f.readlines()
-        sys.stdout.write(lines[0] + '\n')
-        f.seek(0)
-        f.truncate()
-        f.writelines(lines[1:])                  
-except:
-    pass
+        lines = f.readines()
+        if len(lines) != 0:
+            sys.stdout.write(lines[0] + '\n')
+            f.seek(0)
+            f.truncate()
+            f.writelines(lines[1:])
+        elif len(lines) == 0:
+            print('No alerts to process.')
+        else:
+            print('Unknown file length.')
+else:
+    print('No alert file to parse.')
